Move common Ansible args decoding to module_utils.
Now that we are a collection we can implement the original
idea of having the decoding of ansible vars in a common place.
