+++ plugins/action/ci_script.py
@@ -95,15 +95,14 @@ import glob
 import pathlib
 import re
 import uuid
-import yaml
-import typing
 
 
 from ansible.plugins.action import ActionBase
 from ansible.errors import AnsibleActionFail
 
-from ansible.parsing.yaml.dumper import AnsibleDumper
-from ansible.utils.unsafe_proxy import AnsibleUnsafeText, AnsibleUnsafeBytes
+from ansible_collections.cifmw.general.plugins.module_utils.encoding import (
+    ansible_encoding,
+)
 
 TMPL_SCRIPT = """#!/bin/bash
 set -euo pipefail
@@ -113,38 +112,6 @@ exec > >(tee -i %(logpath)s) 2>&1
 %(content)s
 %(popcmd)s
 """
-
-
-def decode_ansible_raw(data: typing.Any) -> typing.Any:
-    """Converts an Ansible var to a python native one
-
-    Ansible raw input args can contain AnsibleUnicodes or AnsibleUnsafes
-    that are not intended to be manipulated directly.
-    This function converts the given variable to a one that only contains
-    python built-in types.
-
-    Args:
-        data: The usafe Ansible content to decode
-
-    Returns: The python types based result
-
-    """
-    if isinstance(data, list):
-        return [decode_ansible_raw(_data) for _data in data]
-    elif isinstance(data, tuple):
-        return tuple(decode_ansible_raw(_data) for _data in data)
-    if isinstance(data, dict):
-        return yaml.load(
-            yaml.dump(
-                data, Dumper=AnsibleDumper, default_flow_style=False, allow_unicode=True
-            ),
-            Loader=yaml.Loader,
-        )
-    if isinstance(data, AnsibleUnsafeText):
-        return str(data)
-    if isinstance(data, AnsibleUnsafeBytes):
-        return bytes(data)
-    return data
 
 
 class ActionModule(ActionBase):
@@ -170,7 +137,7 @@ class ActionModule(ActionBase):
     def run(self, tmp=None, task_vars=None):
         super(ActionModule, self).run(tmp, task_vars)
 
-        task_args = decode_ansible_raw(self._task.args)
+        task_args = ansible_encoding.decode_ansible_raw(self._task.args)
         if "output_dir" not in task_args:
             raise AnsibleActionFail("output_dir parameter is missing")
